refactor(ui): decompose GeneralTab and eliminate +page.svelte prop drilling (#656)
